+++ src/main/java/umc/meme/shop/domain/artist/controller/ArtistController.java
@@ -23,27 +23,6 @@ public class ArtistController {
         return ApiResponse.SuccessResponse(SuccessStatus.PROFILE_UPDATE);
     }
 
-    @Operation(summary = "포트폴리오 조회", description = "포트폴리오를 조회하는 API입니다.")
-    @GetMapping("/mypage/{userId}/portfolio")
-    public ApiResponse getPortfolio(@PathVariable Long userId){
-        return ApiResponse.SuccessResponse(SuccessStatus.PORTFOLIO_GET, "");
-    }
-
-    @Operation(summary = "포트폴리오 생성", description = "포트폴리오를 생성하는 API입니다.")
-    @PostMapping("/mypage/{userId}/portfolio")
-    public ApiResponse createPortfolio(@PathVariable Long userId, @RequestBody CreatePortfolioDto portfolioDto){
-        // TODO: PortfolioImg 추가
-        artistService.createPortfolio(userId, portfolioDto);
-        return ApiResponse.SuccessResponse(SuccessStatus.PORTFOLIO_CREATE);
-    }
-
-    @Operation(summary = "포트폴리오 수정/삭제", description = "포트폴리오를 조회하는 API입니다.")
-    @PatchMapping("/mypage/{userId}/portfolio")
-    public ApiResponse updatePortfolio(@PathVariable Long userId, @RequestBody UpdatePortfolioDto portfolioDto){
-        // TODO: PortfolioImg 추가
-        return ApiResponse.SuccessResponse(SuccessStatus.PORTFOLIO_UPDATE);
-    }
-
     @Operation(summary = "예약 조회", description = "예약 정보를 조회하는 API입니다.")
     @GetMapping("/mypage/{userId}/reservation")
     public ApiResponse getReservation(@PathVariable Long userId){
