+++ src/main/java/umc/meme/shop/domain/portfolio/service/PortfolioService.java
@@ -1,4 +1,4 @@
-package umc.meme.shop.domain.artist.service;
+package umc.meme.shop.domain.portfolio.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -8,30 +8,28 @@ import umc.meme.shop.domain.artist.repository.ArtistRepository;
 import umc.meme.shop.domain.portfolio.dto.request.CreatePortfolioDto;
 import umc.meme.shop.domain.portfolio.entity.Portfolio;
 import umc.meme.shop.domain.portfolio.repository.PortfolioRepository;
-import umc.meme.shop.domain.reservation.dto.response.ReservationResponseDto;
-import umc.meme.shop.domain.reservation.entity.Reservation;
-import umc.meme.shop.domain.reservation.repository.ReservationRepository;
 import umc.meme.shop.global.ErrorStatus;
 import umc.meme.shop.global.exception.GlobalException;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
 @Service
 @RequiredArgsConstructor
-public class ArtistService {
+public class PortfolioService {
     private final ArtistRepository artistRepository;
-    private final ReservationRepository reservationRepository;
+    private final PortfolioRepository portfolioRepository;
 
-    //예약 조회
-    public List<ReservationResponseDto> getReservation(Long artistId){
+    //포트폴리오 생성
+    @Transactional
+    public void createPortfolio(Long artistId, CreatePortfolioDto portfolioDto){
         Artist artist = artistRepository.findById(artistId)
                 .orElseThrow(() -> new GlobalException(ErrorStatus.NOT_EXIST_USER));
 
-         List<Reservation> reservationList = reservationRepository.findByArtist(artist);
-         return reservationList.stream()
-                 .map(ReservationResponseDto::from)
-                 .collect(Collectors.toList());
+        Portfolio portfolio = Portfolio.builder()
+                .artist(artist)
+                .category(portfolioDto.getCategory())
+                .makeupName(portfolioDto.getMakeupName())
+                .info(portfolioDto.getInfo())
+                .price(portfolioDto.getPrice())
+                .build();
+        portfolioRepository.save(portfolio);
     }
 }
